hw8. refactoring to app->run
